Move glob expansion to new module
