+++ src/expand.rs
@@ -1,0 +1,58 @@
+use crate::error::{Error, Result};
+use crate::manifest::Name;
+use crate::Test;
+use std::path::PathBuf;
+
+#[derive(Debug)]
+pub(crate) struct ExpandedTest {
+    pub name: Name,
+    pub test: Test,
+    pub error: Option<Error>,
+}
+
+pub(crate) fn expand_globs(tests: &[Test]) -> Vec<ExpandedTest> {
+    fn glob(pattern: &str) -> Result<Vec<PathBuf>> {
+        let mut paths = glob::glob(pattern)?
+            .map(|entry| entry.map_err(Error::from))
+            .collect::<Result<Vec<PathBuf>>>()?;
+        paths.sort();
+        Ok(paths)
+    }
+
+    fn bin_name(i: usize) -> Name {
+        Name(format!("trybuild{:03}", i))
+    }
+
+    let mut vec = Vec::new();
+
+    for test in tests {
+        let mut expanded = ExpandedTest {
+            name: bin_name(vec.len()),
+            test: test.clone(),
+            error: None,
+        };
+        if let Some(utf8) = test.path.to_str() {
+            if utf8.contains('*') {
+                match glob(utf8) {
+                    Ok(paths) => {
+                        for path in paths {
+                            vec.push(ExpandedTest {
+                                name: bin_name(vec.len()),
+                                test: Test {
+                                    path,
+                                    expected: expanded.test.expected,
+                                },
+                                error: None,
+                            });
+                        }
+                        continue;
+                    }
+                    Err(error) => expanded.error = Some(error),
+                }
+            }
+        }
+        vec.push(expanded);
+    }
+
+    vec
+}
