+++ src/run.rs
@@ -3,6 +3,7 @@ use crate::dependencies::{self, Dependency};
 use crate::directory::Directory;
 use crate::env::Update;
 use crate::error::{Error, Result};
+use crate::expand::{expand_globs, ExpandedTest};
 use crate::flock::Lock;
 use crate::manifest::{Bin, Build, Config, Manifest, Name, Package, Workspace};
 use crate::message::{self, Fail, Warn};
@@ -469,60 +470,6 @@ fn check_exists(path: &Path) -> Result<()> {
         Ok(_) => Ok(()),
         Err(err) => Err(Error::Open(path.to_owned(), err)),
     }
-}
-
-#[derive(Debug)]
-struct ExpandedTest {
-    name: Name,
-    test: Test,
-    error: Option<Error>,
-}
-
-fn expand_globs(tests: &[Test]) -> Vec<ExpandedTest> {
-    fn glob(pattern: &str) -> Result<Vec<PathBuf>> {
-        let mut paths = glob::glob(pattern)?
-            .map(|entry| entry.map_err(Error::from))
-            .collect::<Result<Vec<PathBuf>>>()?;
-        paths.sort();
-        Ok(paths)
-    }
-
-    fn bin_name(i: usize) -> Name {
-        Name(format!("trybuild{:03}", i))
-    }
-
-    let mut vec = Vec::new();
-
-    for test in tests {
-        let mut expanded = ExpandedTest {
-            name: bin_name(vec.len()),
-            test: test.clone(),
-            error: None,
-        };
-        if let Some(utf8) = test.path.to_str() {
-            if utf8.contains('*') {
-                match glob(utf8) {
-                    Ok(paths) => {
-                        for path in paths {
-                            vec.push(ExpandedTest {
-                                name: bin_name(vec.len()),
-                                test: Test {
-                                    path,
-                                    expected: expanded.test.expected,
-                                },
-                                error: None,
-                            });
-                        }
-                        continue;
-                    }
-                    Err(error) => expanded.error = Some(error),
-                }
-            }
-        }
-        vec.push(expanded);
-    }
-
-    vec
 }
 
 impl ExpandedTest {
